Future Plan

This is a wrapper function around doFuture::registerDoFuture() foreach::registerDoSEQ(), and future::plan(). The strategy is the evaluation function to use in resolving a future. This function quickly enables the user to toggle between sequential and multiprocess strategies with an option to set a custom strategy when needed. This wrapper function will continue to evolve to adjust to my needs.

future_plan(par = FALSE, foreach_seq = TRUE, ...)

Arguments

par
Logical, or strategy. If par = TRUE, sets multiprocess future plan. In this case, if multicore evaluation is supported, it will be used, otherwise multisession is used. If par = FALSE, sets sequential strategy. If par is NOT logical, par is passed as the strategy argument to future::plan().

foreach_seq
Logical. If TRUE, use foreach::registerDoSEQ() for sequential strategy. If FALSE, use sequential strategy in future::plan().

...
Other arguments (other than strategy) to pass to future::plan().

Value

Nothing.

See also

Examples

library(future)
# set `sequential` strategy--------------------------------------------------
future_plan(par = FALSE)
plan(strategy = NULL)

#> sequential:
#> - args: function (expr, envir = parent.frame(), substitute = TRUE, lazy = FALSE, seed = NULL, globals = TRUE, local = TRUE, earlySignal = FALSE, label = NULL, ...)
#> - tweaked: FALSE
#> - call: NULL

# set `multiprocess` strategy------------------------------------------------
future_plan(par = TRUE)
plan(strategy = NULL)

#> multiprocess:
#> - args: function (expr, envir = parent.frame(), substitute = TRUE, lazy = FALSE, seed = NULL, globals = TRUE, workers = availableCores(), gc = FALSE, earlySignal = FALSE, label = NULL, ...)
#> - tweaked: FALSE
#> - call: plan(strategy = strategy, ...)

# set `multisession` strategy------------------------------------------------
future_plan(par = multisession)
plan(strategy = NULL)

#> multisession:
#> - args: function (expr, envir = parent.frame(), substitute = TRUE, lazy = FALSE, seed = NULL, globals = TRUE, persistent = FALSE, workers = availableCores(), gc = FALSE, earlySignal = FALSE, label = NULL, ...)
#> - tweaked: FALSE
#> - call: plan(strategy = strategy, ...)

# set `sequential` strategy--------------------------------------------------
future_plan(par = FALSE, foreach_seq = FALSE)
plan(strategy = NULL)

#> sequential:
#> - args: function (expr, envir = parent.frame(), substitute = TRUE, lazy = FALSE, seed = NULL, globals = TRUE, local = TRUE, earlySignal = FALSE, label = NULL, ...)
#> - tweaked: FALSE
#> - call: plan(strategy = strategy, ...)