Parallel lapply

This is a wrapper function for different implementations of lapply.

par_lapply(
  X,
  FUN,
  ...,
  par = TRUE,
  ncores = NULL,
  mc = TRUE,
  lb = FALSE,
  cl_eval = FALSE,
  cl_export = FALSE,
  cl_expr,
  cl_vars,
  rbind = NULL
)

Arguments

X
Vector or list. Each element will be passed as the first argument of FUN.

FUN
Function to be applied to each element of X.

...
Arguments to pass to FUN.

par
Logical. If TRUE, use multiple cores. If FALSE, use lapply().

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

mc
Logical. If TRUE, use parallel::mclapply(). If FALSE, use parallel::parLapply() or parallel::parLapplyLB(). Ignored if par = FALSE.

lb
Logical. If TRUE use parallel::parLapplyLB(). If FALSE, use parallel::parLapply(). Ignored if par = FALSE and mc = TRUE.

cl_eval
Logical. Execute parallel::clusterEvalQ() using cl_expr. Ignored if mc = TRUE.

cl_export
Logical. Execute parallel::clusterExport() using cl_vars. Ignored if mc = TRUE.

cl_expr
Expression. Expression passed to parallel::clusterEvalQ() Ignored if mc = TRUE.

cl_vars
Character vector. Names of objects to pass to parallel::clusterExport() Ignored if mc = TRUE.

rbind
NULL or logical. If rbind = NULL, returns the list produced. If TRUE, uses rbind() to bind the rows of the list produced. If FALSE, uses cbind() to bind the columns of the list produced. Test that each element of the output list has the appropriate dimensions for binding before using this option.

Value

Returns the results of FUN for each element of X.

See also

Examples

X <- rep(x = 5, times = 5)
# rbind = NULL----------------------------------------------------------------
set.seed(42)
par_lapply(
  X = X,
  FUN = rnorm,
  par = FALSE
)

#> [[1]]
#> [1]  1.3709584 -0.5646982  0.3631284  0.6328626  0.4042683
#> 
#> [[2]]
#> [1] -0.10612452  1.51152200 -0.09465904  2.01842371 -0.06271410
#> 
#> [[3]]
#> [1]  1.3048697  2.2866454 -1.3888607 -0.2787888 -0.1333213
#> 
#> [[4]]
#> [1]  0.6359504 -0.2842529 -2.6564554 -2.4404669  1.3201133
#> 
#> [[5]]
#> [1] -0.3066386 -1.7813084 -0.1719174  1.2146747  1.8951935
#> 

#-----------------------------------------------------------------------------
# rbind = TRUE----------------------------------------------------------------
set.seed(42)
par_lapply(
  X = X,
  FUN = rnorm,
  par = FALSE,
  rbind = TRUE
)

#>            [,1]       [,2]        [,3]       [,4]       [,5]
#> [1,]  1.3709584 -0.5646982  0.36312841  0.6328626  0.4042683
#> [2,] -0.1061245  1.5115220 -0.09465904  2.0184237 -0.0627141
#> [3,]  1.3048697  2.2866454 -1.38886070 -0.2787888 -0.1333213
#> [4,]  0.6359504 -0.2842529 -2.65645542 -2.4404669  1.3201133
#> [5,] -0.3066386 -1.7813084 -0.17191736  1.2146747  1.8951935

#-----------------------------------------------------------------------------
# rbind = FALSE---------------------------------------------------------------
set.seed(42)
par_lapply(
  X = X,
  FUN = rnorm,
  par = FALSE,
  rbind = FALSE
)

#>            [,1]        [,2]       [,3]       [,4]       [,5]
#> [1,]  1.3709584 -0.10612452  1.3048697  0.6359504 -0.3066386
#> [2,] -0.5646982  1.51152200  2.2866454 -0.2842529 -1.7813084
#> [3,]  0.3631284 -0.09465904 -1.3888607 -2.6564554 -0.1719174
#> [4,]  0.6328626  2.01842371 -0.2787888 -2.4404669  1.2146747
#> [5,]  0.4042683 -0.06271410 -0.1333213  1.3201133  1.8951935

#-----------------------------------------------------------------------------